+++ lib/rails_uuid_pk/mysql2_adapter_extension.rb
@@ -1,0 +1,9 @@
+module RailsUuidPk
+  module Mysql2AdapterExtension
+    def native_database_types
+      super.merge(
+        uuid: { name: "varchar", limit: 36 }
+      )
+    end
+  end
+end